[refactor] Refactor search component
