+++ src/components/Search/index.tsx
@@ -1,36 +1,14 @@
-import { useEffect, useState } from "react";
-import { useDashboardSettings } from "../../context/DashboardSettingsContext";
-// import useGoogleSuggestions from "../../hooks/search/useGoogleSuggestions";
+import { useSearch } from "../../hooks/search";
 
 const Search = (): React.JSX.Element => {
-  const { isSearchToggled } = useDashboardSettings();
-  // const { suggestions } = useGoogleSuggestions(query);
-  const [query, setQuery] = useState("");
-  const [shouldRender, setShouldRender] = useState(isSearchToggled);
-
-  useEffect(() => {
-    if (isSearchToggled) setShouldRender(true);
-  }, [isSearchToggled]);
-
-  const handleAnimationEnd = () => {
-    if (!isSearchToggled) setShouldRender(false);
-  };
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    if (!query.trim()) return;
-
-    const baseUrl = "https://www.google.com/search";
-
-    window.location.href = `${baseUrl}?q=${encodeURIComponent(query)}`;
-  };
-
-  // const handleSuggestionClick = (suggestion: string) => {
-  //   setQuery(suggestion);
-  //   const baseUrl = "https://www.google.com/search";
-
-  //   window.location.href = `${baseUrl}?q=${encodeURIComponent(suggestion)}`;
-  // };
+  const {
+    query,
+    setQuery,
+    handleAnimationEnd,
+    handleSubmit,
+    shouldRender,
+    isSearchToggled,
+  } = useSearch();
 
   if (!shouldRender) return <></>;
 
@@ -59,21 +37,6 @@ const Search = (): React.JSX.Element => {
           Search
         </button>
       </div>
-
-      {/* TODO suggestions has CORS issue upon building */}
-      {/* {suggestions.length > 0 && (
-        <ul className="absolute z-10 w-full bg-white dark:bg-background-dark-secondary border mt-1 rounded shadow">
-          {suggestions.map((s, idx) => (
-            <li
-              key={idx}
-              onClick={() => handleSuggestionClick(s)}
-              className="px-4 py-2 hover:bg-gray-100 cursor-pointer"
-            >
-              {s}
-            </li>
-          ))}
-        </ul>
-      )} */}
     </form>
   );
 };
